Remove Groovy dependency from commons
