+++ dsl/src/main/java/com/powsybl/dsl/PowsyblDslAstTransformation.java
@@ -6,20 +6,66 @@
  */
 package com.powsybl.dsl;
 
-import com.powsybl.commons.ast.AbstractAstTransformation;
+import groovy.inspect.swingui.AstNodeToScriptVisitor;
 import org.codehaus.groovy.ast.ASTNode;
 import org.codehaus.groovy.ast.ClassCodeExpressionTransformer;
+import org.codehaus.groovy.ast.MethodNode;
+import org.codehaus.groovy.ast.ModuleNode;
 import org.codehaus.groovy.ast.expr.*;
+import org.codehaus.groovy.ast.stmt.BlockStatement;
 import org.codehaus.groovy.control.SourceUnit;
+import org.codehaus.groovy.transform.ASTTransformation;
 import org.codehaus.groovy.transform.GroovyASTTransformation;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.io.IOException;
+import java.io.StringWriter;
+import java.io.UncheckedIOException;
+import java.util.List;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
 @GroovyASTTransformation
-public class PowsyblDslAstTransformation extends AbstractAstTransformation {
+public class PowsyblDslAstTransformation implements ASTTransformation {
+
+    protected static final Logger LOGGER = LoggerFactory.getLogger(PowsyblDslAstTransformation.class);
 
     private static final boolean DEBUG = false;
+
+    private static void printAST(BlockStatement blockStatement) {
+        try (StringWriter writer = new StringWriter()) {
+            blockStatement.visit(new AstNodeToScriptVisitor(writer));
+            writer.flush();
+            if (LOGGER.isDebugEnabled()) {
+                LOGGER.debug(writer.toString());
+            }
+        } catch (IOException e) {
+            throw new UncheckedIOException(e);
+        }
+    }
+
+    protected void visit(SourceUnit sourceUnit, ClassCodeExpressionTransformer transformer, boolean debug) {
+        LOGGER.trace("Apply AST transformation");
+        ModuleNode ast = sourceUnit.getAST();
+        BlockStatement blockStatement = ast.getStatementBlock();
+
+        if (debug) {
+            printAST(blockStatement);
+        }
+
+        List<MethodNode> methods = ast.getMethods();
+        for (MethodNode methodNode : methods) {
+            methodNode.getCode().visit(transformer);
+        }
+
+        blockStatement.visit(transformer);
+
+        if (debug) {
+            printAST(blockStatement);
+        }
+    }
 
     public void visit(ASTNode[] nodes, SourceUnit sourceUnit) {
         visit(sourceUnit, new CustomClassCodeExpressionTransformer(sourceUnit), DEBUG);
